+++ react/hooks/useOrganization.ts
@@ -3,14 +3,18 @@ import { useQuery } from 'react-apollo'
 import type { QueryGetUsersArgs } from 'vtex.b2b-organizations-graphql'
 import type { SessionSuccess } from 'vtex.session-client'
 import { useFullSession } from 'vtex.session-client'
+import { Query } from 'vtex.storefront-permissions'
 
 import GET_ORGANIZATION from '../graphql/getOrganization.graphql'
+import GET_PERMISSIONS from '../graphql/getPermissions.graphql'
 import type {
   CustomOrganization,
   GetOrganizationQuery,
   SessionOrganizationData,
 } from '../typings'
 import { NAMESPACE_ITEMS, SESSION_NAMESPACE } from '../utils'
+
+type GetPermissionsQuery = Pick<Query, 'checkUserPermission'>
 
 export function useOrganization() {
   const { data: sessionData, loading: sessionLoading } = useFullSession({
@@ -35,15 +39,26 @@ export function useOrganization() {
     variables: { costCenterId, organizationId },
   })
 
+  const {
+    data: permissionsData,
+    loading: permissionsLoading,
+  } = useQuery<GetPermissionsQuery>(GET_PERMISSIONS, {
+    ssr: false,
+  })
+
   const organization: CustomOrganization = useMemo(() => {
     return {
       ...data?.getOrganizationByIdStorefront,
       users: data?.getUsers,
       costCenter: data?.getCostCenterByIdStorefront,
+      role: permissionsData?.checkUserPermission?.role?.id ?? '',
     }
-  }, [data])
+  }, [data, permissionsData])
 
-  const loading = sessionLoading || organizationLoading
+  const loading = useMemo(
+    () => sessionLoading || organizationLoading || permissionsLoading,
+    [organizationLoading, permissionsLoading, sessionLoading]
+  )
 
   return { organization, loading }
 }